% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{compute.ClosureAggregates}
\alias{compute.ClosureAggregates}
\title{Closure Aggregates based on Floor Date}
\usage{
compute.ClosureAggregates(tib_df, col_dt = "FloorDate", col_week = "Week")
}
\arguments{
\item{tib_df}{Tibble with Week and Floor Date Information}

\item{col_dt}{Date to be aggregated on}

\item{col_week}{Week to be aggregated on}
}
\value{
tibble with Week, FloorDate and NumClosed Columns
}
\description{
takes a tibble with a Week Column and a Floor date Column and then returns an aggregated tibble
}
