% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{compute.CycleTime}
\alias{compute.CycleTime}
\title{Adds a column to the tibble with the cycle time and returns a tibble}
\usage{
compute.CycleTime(tib_df, col_created_on = "crdt", col_closed_date = "cldt")
}
\arguments{
\item{tib_df}{Tibble containing the created date and closed date}

\item{col_created_on}{String which has the name of the created_on column defaults to "crdt"}

\item{col_closed_date}{String which has the name of the closed_date column defaults to "cldt"}
}
\value{
A tibble with a new column called "cylt" having the cycle time
}
\description{
Adds a column to the input tibble with the calculated Cycle Time
Accepts a tibble which has the created date and the closed date as the
columns and calculates the cycle time from it and adds it as another
column.
}
\seealso{
getClosedCases()
}
