% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{compute.FloorDateBased.ClosureAggregates}
\alias{compute.FloorDateBased.ClosureAggregates}
\title{Computes closure aggregates for every FloorDate}
\usage{
compute.FloorDateBased.ClosureAggregates(tib_df, col_dt = "FloorDate")
}
\arguments{
\item{tib_df}{Unaggregated tibble  with the week number as a variable}

\item{col_dt}{A column which has the FloorDate as the variable}
}
\value{
tib_df the aggregated tibble
}
\description{
Computes the Closure aggregates for every week. Grouping is
done for every Floor Date followed by summarisation of the
Number closed
}
