% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{compute.NumWeeksForClosureCount}
\alias{compute.NumWeeksForClosureCount}
\title{Computes the number of weeks with a particular closure count}
\usage{
compute.NumWeeksForClosureCount(tib_df, col_numclosed = "NumClosed")
}
\arguments{
\item{tib_df}{The Tibble which has "Number of Weeks" for each "Closure Count}

\item{col_numclosed}{The column name of the tibble which has the
number of closed items}
}
\description{
Can be used for finding out how predictable the delivery is
}
\details{
get.FilteredTibble(fileName="data/example_delivery.csv") %>%  compute.CycleTime() %>% compute.Week() %>% compute.ClosureAggregates() %>% compute.NumWeeksForClosureCount()
}
