% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{compute.PriorityBased.ClosureAggregates}
\alias{compute.PriorityBased.ClosureAggregates}
\title{Generates a tibble with the Date, Priority and Number of Work Items
Closed}
\usage{
compute.PriorityBased.ClosureAggregates(
  tib_df,
  col_date = "FloorDate",
  col_priority = "Priority"
)
}
\arguments{
\item{tib_df}{Tibble containing complete cases}

\item{col_date}{Column Name of with the Floor Date derived from
the Closed Date}

\item{col_priority}{Column name of the priority of the closed items}
}
\description{
Receives a tibble with complete cases (closed cases) and then
generates a tibble with the Date, Priority and Number of Work Items
Closed
}
\details{
get.FilteredTibble(fileName="data/example_delivery.csv") %>%  get.ClosedCases() %>% compute.CycleTime() %>% compute.Week() %>% compute.PriorityBased.ClosureAggregates()
}
\seealso{
compute.Week()
}
