% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{compute.Week}
\alias{compute.Week}
\title{Adds a week number to the tibble and returns}
\usage{
compute.Week(tib_df, col_closed_date = "cldt")
}
\arguments{
\item{tib_df}{Input tibble with the Closed Date}

\item{col_closed_date}{The Column name which has the closed date}
}
\description{
Adds a column called Week using the the Floor Date
derived from Closed date
}
