% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{compute.Weekly.ClosureAggregates}
\alias{compute.Weekly.ClosureAggregates}
\title{Computes closure aggregates for every week}
\usage{
compute.Weekly.ClosureAggregates(tib_df, col_week = "Week")
}
\arguments{
\item{tib_df}{Unaggregated tibble  with the week number as a variable}

\item{col_week}{A column which has the week number as the variable}
}
\value{
tib_df the aggregated tibble
}
\description{
Computes the Closure aggregates for every week. Grouping is
done for every week followed by summarisation of the Number closed
}
