% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{compute.WIP}
\alias{compute.WIP}
\title{Takes a tibble and returns a tibble with wip for each date}
\usage{
compute.WIP(tib, col_created_date = "crdt", col_closed_date = "cldt")
}
\arguments{
\item{tib}{Tibble containing Open and Closed Dates.}

\item{col_created_date}{Column name with created date (default: crdt)}

\item{col_closed_date}{Column name with closed date (default: cldt)}
}
\value{
Tibble with Date (Created Date) and Sum of WIP In Days
}
\description{
Takes a tibble and returns a tibble with wip for each date
}
