% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{constuct.WIPTibble}
\alias{constuct.WIPTibble}
\alias{construct.WIPTibble}
\title{Constructs a tibble based on a data file, date from and type to be used for computing the WIP}
\usage{
construct.WIPTibble(tib)
}
\arguments{
\item{tib}{tibble to compute the WiP for every day}
}
\value{
A tibble
}
\description{
Constructs a tibble based on various parameters
}
