% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{exclude.OpenCases}
\alias{exclude.OpenCases}
\title{Alias for get.ClosedCases()}
\usage{
exclude.OpenCases(tib_df, col_closed_date = "cldt")
}
\arguments{
\item{tib_df}{Input data frame which has unclosed cases as well}

\item{col_closed_date}{Column which has the closed date}
}
\description{
Filters out all cases other than closed cases
and Alias of get.ClosedCases
}
