% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.Age.In.Days}
\alias{get.Age.In.Days}
\title{gets the age in days from the begining and closing date}
\usage{
get.Age.In.Days(beginDate, endDate = Sys.Date())
}
\arguments{
\item{beginDate}{the begining date}

\item{endDate}{defaults to system date}
}
\description{
Gets the age of the work item in days
}
\seealso{
compute.CycleTime
}
