% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{get.ClosedCases}
\alias{get.ClosedCases}
\title{Takes a tibble and returns a list of closed cases}
\usage{
get.ClosedCases(tib_df, col_closed_date = "cldt")
}
\arguments{
\item{tib_df}{Input data frame which has unclosed cases as well}

\item{col_closed_date}{Column which has the closed date}
}
\description{
Filters out all cases other than closed cases
}
\details{
get.FilteredTibble(fileName="example_delivery.csv") %>% get.ClosedCases()
}
\seealso{
compute.CycleTime
}
