% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{get.FilteredTibble}
\alias{get.FilteredTibble}
\title{get.FilteredTibble(fileName="../Datasets/Opx_2021-06-17_Delivery.csv") %>%   addColCycleTime() %>% computeWeeklyFloorDates() %>% compute.Week()
get.FilteredTibble(fileName="../Datasets/Opx_2021-06-17_Delivery.csv") %>%  addColCycleTime() %>% computeWeeklyFloorDates() %>% computePriorityBasedWeeklyClosureAggregates()
Generate a tibble from a file filtered on a date
Accepts a data file appropriately formatted which contains the Created Date,
closed date, updated date of jira issues and a cut off date and returns
a tibble.
#'
get.FilteredTibble(fileName="data/example_delivery.csv")
get.FilteredTibble(fileName="data/example_delivery.csv", sep='\\t')
get.FilteredTibble(fileName="data/example_delivery.csv", sep=",", date_from=as.Date("2021-03-03"))}
\usage{
get.FilteredTibble(
  fileName,
  sep = "\\t",
  date_from = Sys.Date() - months(6),
  col_created_on = "Created",
  col_updated_date = "Updated",
  col_closed_date = "Closed"
)
}
\arguments{
\item{fileName}{Input file name}

\item{sep}{seperator}

\item{date_from}{cut off date for the data frame}

\item{col_created_on}{Column name of the created column}

\item{col_updated_date}{Column name of the updated column}

\item{col_closed_date}{Column name of the closed date column}
}
\value{
tib A tibble filtered on the cut off date. Closed Date is >= date_from
}
\description{
get.FilteredTibble(fileName="../Datasets/Opx_2021-06-17_Delivery.csv") %>%   addColCycleTime() %>% computeWeeklyFloorDates() %>% compute.Week()
get.FilteredTibble(fileName="../Datasets/Opx_2021-06-17_Delivery.csv") %>%  addColCycleTime() %>% computeWeeklyFloorDates() %>% computePriorityBasedWeeklyClosureAggregates()
Generate a tibble from a file filtered on a date
Accepts a data file appropriately formatted which contains the Created Date,
closed date, updated date of jira issues and a cut off date and returns
a tibble.
#'
get.FilteredTibble(fileName="data/example_delivery.csv")
get.FilteredTibble(fileName="data/example_delivery.csv", sep='\\t')
get.FilteredTibble(fileName="data/example_delivery.csv", sep=",", date_from=as.Date("2021-03-03"))
}
\author{
"Rajiv Gangadharan"
}
