% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamanip.R
\name{get.WIPInDays}
\alias{get.WIPInDays}
\title{Returns the WIP at a particular date for a created and closed date}
\usage{
get.WIPInDays(created, closed, loop_date = Sys.Date())
}
\arguments{
\item{created}{The Created Date}

\item{closed}{The Closed Date}

\item{loop_date}{The date on which the age in days have to be computed}
}
\value{
Returns a numeric age in days
}
\description{
Accepts the created and closed dates and returns the age in
days for a particular date which is provided.
}
