% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAgeInDays}
\alias{getAgeInDays}
\title{gets the age in days from the begining and closing date}
\usage{
getAgeInDays(beginDate, endDate = Sys.Date())
}
\arguments{
\item{beginDate}{the begining date}

\item{endDate}{defaults to system date}
}
\description{
Gets the age of the work item in days
}
\details{
Tested using
tib$WIPDAYS_2021_02_21 <- apply(tib, 1, getWIPInDays, ymd("2021-02-21"))
Tested using tib$WIP_2021_02_21 <- apply(tib, 1, isWIP, ymd("2021-02-21"))
}
\seealso{
compute.CycleTime
}
