% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.none}
\alias{is.none}
\title{Check if the parameter is None or not}
\usage{
is.none(x)
}
\arguments{
\item{x}{variable for which None value has to be handled}
}
\value{
TRUE or FALSE based of if the Value is None
}
\description{
Purpose is to handle None values inserted into the data files due to python scripts
}
