% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tib_complete_cases}
\alias{tib_complete_cases}
\title{Closed Cases Data}
\format{
A tibble with the variables
\describe{
\item{Key}{Issue ID from JIRA}
\item{Type}{}
\item{Status}{}
\item{Priority}{}
\item{Created}{Data on which created}
\item{Updated}{Updated date}
\item{Closed}{Closed Date}
\item{crdt}{Formatted Created Date}
\item{updt}{Formatted Updated Date}
\item{cldt}{Formatted Closed Date}
}
}
\source{
Exported as tsv from jira.
}
\usage{
tib_complete_cases
}
\description{
This is output from get.ClosedCases() and gives the complete cases for
issues. It drops the rows having NA in the Closed column
}
\keyword{datasets}
