% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toDate}
\alias{toDate}
\title{Takes a string and then converts into date}
\usage{
toDate(argString, dateFormat = "\%Y-\%m-\%d")
}
\arguments{
\item{argString}{A string to convert into date}

\item{dateFormat}{A valid date format}
}
\value{
The string converted into a date
}
\description{
Takes a string and then converts into date, utility function to be used in datamanip.R by many functions.
}
\details{
Convert string into date


kes a string and then converts into date, utility function to be used in datamanip.R by many functions.
}
